+++ src/score2dx/Core/Core.hpp
@@ -90,8 +90,17 @@ public:
         FindAnalysis(const std::string &iidxId)
         const;
 
-        //'' temp test
-        Analyzer & GetAnalyzer() { return mAnalyzer; }
+    //! @note Requires manually call analyze because need date time range, does not auto
+    //! analyze upon LoadDirectory.
+        void
+        AnalyzeActivity(const std::string &iidxId,
+                        const std::string &beginDateTime,
+                        const std::string &endDateTime);
+
+    //! @brief Find if player of IIDX ID has activity analysis.
+        const ActivityAnalysis*
+        FindActivityAnalysis(const std::string &iidxId)
+        const;
 
 private:
     MusicDatabase mMusicDatabase;
@@ -103,8 +112,10 @@ private:
     std::map<std::string, std::map<PlayStyle, std::map<std::string, std::unique_ptr<Csv>>>> mPlayerCsvs;
 
     Analyzer mAnalyzer;
-    //! @brief Map of {IidxId, LastScoreAnalysis}.
+    //! @brief Map of {IidxId, ScoreAnalysis}.
     std::map<std::string, ScoreAnalysis> mPlayerAnalyses;
+    //! @brief Map of {IidxId, ActivityAnalysis}.
+    std::map<std::string, ActivityAnalysis> mPlayerActivityAnalyses;
 
         void
         CreatePlayer(const std::string &iidxId);